requirements.txt
pymupdf
pymupdf4llm
openai
python-dotenv
